Refactor test assertions for clarity and consistency
- Updated assertions in GraphQL schema template tests to improve readability by formatting multiline assertions.
- Simplified test binary JSON line parsing in `xtask/src/main.rs` for better clarity and maintainability.
- Enhanced test cases to ensure consistent formatting and improved readability of assertions across various tests.
