+++ bitloops/src/host/capability_host/runtime_contexts/local_gateways.rs
@@ -1,0 +1,38 @@
+use anyhow::Result;
+use serde_json::Value;
+
+use crate::host::capability_host::gateways::{
+    CanonicalGraphGateway, ProvenanceBuilder, StoreHealthGateway,
+};
+
+pub struct LocalCanonicalGraphGateway;
+
+impl CanonicalGraphGateway for LocalCanonicalGraphGateway {}
+
+pub struct DefaultProvenanceBuilder;
+
+impl ProvenanceBuilder for DefaultProvenanceBuilder {
+    fn build(&self, capability_id: &str, operation: &str, details: Value) -> Value {
+        serde_json::json!({
+            "capability": capability_id,
+            "operation": operation,
+            "details": details,
+        })
+    }
+}
+
+pub struct LocalStoreHealthGateway;
+
+impl StoreHealthGateway for LocalStoreHealthGateway {
+    fn check_relational(&self) -> Result<()> {
+        Ok(())
+    }
+
+    fn check_documents(&self) -> Result<()> {
+        Ok(())
+    }
+
+    fn check_blobs(&self) -> Result<()> {
+        Ok(())
+    }
+}
